wip: refactor File and State
The class File has been refactored into an abstract base class that requires the declaration of an id property.

The State logic has been made more concrete and makes use of the id property of files for automatically determining Deltas. It has been left declared as an ABC because it is convenient to subclass it so that each file system has its own state class.