Release v0.2.0 (#2)
* perf(drive): Make use of the changes API.

Use the changes API to compute deltas for state changes on the Google 
Drive side. Also using the file ID to identify folders

* fix(local): Use inode as folder identifier.

The use of paths for folder identification led to undesired behaviour. 
On Linux, the inode can be used to reliably identify a folder.

* fix(local): create destination directory

When writing to a file, create the destination directory if it doesn't 
exist already.

* fix(io): cement IO exception handling.

The code has been guarded in many places against possible IO exceptions. 
In order to reduce these cases, a wait on deleted files has also been 
introduced.

* fix(fs): enhance abstract FS operations

- Ensure that save states are flushed to disk before shutting down
- Prevent add/remove of the same file when content has changed.